+++ client/src/util/useWriteProps.ts
@@ -1,0 +1,15 @@
+import { createContext, useContext } from 'react';
+
+type WriteType = {
+  title: string;
+  content?: string;
+  tags: string[];
+};
+
+export const WritePropsContext = createContext<WriteType | undefined>(undefined);
+
+export const useWriteProps = () => {
+  const state = useContext(WritePropsContext);
+
+  return state;
+};
